requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
alembic>=1.12.0
pyjwt>=2.8.0
torch>=2.0.0
transformers>=4.35.0
diffusers>=0.24.0
accelerate>=0.24.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
protobuf
rembg>=2.0.0
requests>=2.31.0